Split off usageTable generation into a separate utility
